Project L3 - ESIEE Chatbot
This was my final project for L3-equivalent at ESIEE.
The aim was to make a chatbot for the future students of ESIEE Paris.
It was made with a team of 6 students, including Nicolas AUBRY, Corentin AMARGER, Pierre JANNOT, Lorenzo PINAUD, Clément BRILLET and myself.

The chatbot works using the framework Flask on a webpage not yet published by ESIEE Paris.