RDV Manager Project Android
This was part of the assignement of an Android course at ESIEE Paris.
The details of the subject are shown in a pdf contained in the directory.
The whole project was not entirely finished but the main functionalities are available.
This project was made alongside with Aurélien Argoul, another student from ESIEE Paris.